organizando os testes em pastas e refatorando o teste de integração
