+++ tests/units/session-service.test.js
